OBANYC-373: With @Autowired TDS reference extract and store next stop id and next stop distance.
Using Sheldon's setup of the TransitDataService reference in data-sources.xml, I was able to get the VehicleStatusBean for each incoming vehicleId. Then if the vehicle status
contains a TripStatusBean, I get the stop id of the next stop and the distance to the next stop and stored the values, otherwise I leave the TDS values null. I modified pom.xml
to add profiles for tds.path matching the profiles in the onebusaway-nyc-api-webapp pom.