Fix testActivePulloutsForInvalidBus
Recent 2.4.1 -> 2.8.8 jackson upgrade broke this test, because the behavior about fields that are null when JAXB annotations are getting processed changed. It used to default to NON-NULL, which means that null fields are not serialized unless the annotation says so, but now it uses the dsupplied default behavior, or the global default (which is correct in my opinion). This is discussed here: https://github.com/FasterXML/jackson-module-jaxb-annotations/issues/39