3.1.10 in a nutshell
this commit doesn't include working tests and is breaking because a git nightmare happened while trying to push the last month of work on this (and there was a month's work that hadn't been pushed due to a much more minor git issue that seemed low priority). recovering tests next, but first lets get this off of my laptop.